Successful generation of downloadable pdf
Have successfully implemented the conversion of HTML to PDF for download using the python pdfkit library (https://pypi.python.org/pypi/pdfkit). This library requires the wkhtmltopdf program which can be found here (https://wkhtmltopdf.org/). This program also requires to be added to your $PATH but pdfkit offers a solution to that. Happy coding!!