Introduced the use of Fragments.
Brought in the use of Fragments in the Android app. Fragments are a great tool for UI as they provide flexibility in UI aspects such as rotation and moving through different device screen sizes. However, the app now has a error when taking pictures during registration. Hoping to fix that by the end of today.