Add files via upload
Based on the changes to the structure of the app, this is a revised version of its back end to match the new structure. It is incomplete but the remaining bits should be up soon. To get the latest version of the database, just visit the "alembic.ini" file and change the sqlalchemy.url to the appropriate user name, password and db name then run "alembic upgrade head" in the terminal. I have put in two sqlalchemy urls for MySQL and SQLite. Alternatively, you can can import the "dreamteam.sql" and run it. Remember to also change the URI in the app/__init__.py file. PS: Some of the features of the website will require internet access to work because of the jQuery AJAX requirements